+++ docs/components/form/Radio.mdx
@@ -1,6 +1,8 @@
 import { Meta, Canvas, ArgsTable } from '@storybook/addon-docs';
 import { Radio } from '../../../src/form/Radio';
+import { RadioGroup } from '../../../src/form/Radio';
 import * as RadioStories from '../../../src/form/Radio/Radio.story';
+import * as RadioGroupStories from '../../../src/form/Radio/RadioGroup.story';
 
 <Meta title="Docs/Components/Form/Radio" />
 
@@ -16,6 +18,8 @@ ensuring users choose a single item from a list, providing clear decision-making
 <Canvas of={RadioStories.Sizes} />
 ### Disabled state
 <Canvas of={RadioStories.Disabled} />
+### Group of Radio buttons with a single form control
+<Canvas of={RadioGroupStories.Simple} />
 
 ## Design Tokens
 This component uses the following design tokens:
@@ -28,4 +32,8 @@ This component uses the following design tokens:
 Learn more about design tokens in the [design tokens documentation](?path=/story/docs-design-tokens-getting-started--page).
 
 ## API
+### Radio
 <ArgsTable of={Radio} />
+
+### RadioGroup
+<ArgsTable of={RadioGroup} />
