--- docs/components/form/RadioGroup.mdx
@@ -1,29 +1,0 @@
-import { Meta, Canvas, ArgsTable } from '@storybook/addon-docs';
-import { RadioGroup } from '../../../src/form/RadioGroup';
-import * as RadioGroupStories from '../../../src/form/RadioGroup/RadioGroup.story';
-
-<Meta title="Docs/Components/Form/RadioGroup" />
-
-# RadioGroup
-
-RadioGroup is used to group a set of Radio button elements ensuring users choose a single item from a list, providing clear decision-making and control.
-
-## Examples
-### Vertically aligned
-<Canvas sourceState="shown" of={RadioGroupStories.Vertical} />
-### Horizontally aligned
-<Canvas sourceState="shown" of={RadioGroupStories.Horizontal} />
-
-
-## Design Tokens
-This component uses the following design tokens (same as the [Radio](?path=/docs/docs-components-form-radio--docs) component):
-
-- `radio-label-color`
-- `radio-stroke`
-- `radio-indicator-active`
-- `radio-stroke-active`
-
-Learn more about design tokens in the [design tokens documentation](?path=/story/docs-design-tokens-getting-started--page).
-
-## API
-<ArgsTable of={RadioGroup} />
