+++ src/form/Radio/index.ts
@@ -1,1 +1,2 @@
 export * from './Radio';
+export * from './RadioGroup';
