--- src/form/RadioGroup/index.ts
@@ -1,1 +1,0 @@
-export * from './RadioGroup';
